netlify/functions/requirements.txt
flask==3.0.2
requests==2.32.4
markdown==3.5.2
python-dotenv==1.0.1
openai==1.12.0
google-generativeai==0.3.2
huggingface-hub==0.20.3
